Hostage design, improved support functions
Report on trouble encountered on the way with local install:
#abremaud@esciencefactory.com, 21012016
#Problem: Could not retrieve mongo document via its id field from
Django using mongoDB engine.
#Error msg: invalid literal for int() in base 10
#Diagnostic: conflict between Django expecting integer and MongoDB
using strings for id.
#Diagnostic: Need Django non-rel version installed and order in which
dependencies get installed may matter.
#Diagnostic: Answer by jdi and comments from
http://stackoverflow.com/questions/8861516/object-is-not-getting-saved-w
hile-using-django-non-rel
#Solution: Reinstall virtual environment dependencies from scratch.
(answer by Roberst Moskal from
http://stackoverflow.com/questions/11248073/what-is-the-easiest-way-to-r
emove-all-packages-installed-by-pip)
#Solution: virtualenv -p /usr/bin/python2.7 —clear
NAME_OF_PARTICULAR_VIRTUALENV
#Solution: Start reinstalling according to Setup steps from Django
MongoDB engine at
https://django-mongodb-engine.readthedocs.org/en/latest/topics/setup.htm
l
#Solution: Then pip install requirements.txt, with Django,
django-mongdb-engine and djangotoolobox disabled.

+++ searchapp/mongo_support_functions.py
@@ -1,21 +1,27 @@
-import json
+import json, bson
+
 
 
 from .models import *
 
+
 def DeleteFromPOSTinfo( db_id ):
+    del_flag = False
     try:
-        if len(db_id)>0:
+        if db_id and type(db_id) is unicode and len(db_id)>0:
             Annotation.objects.get(id=db_id).delete()
+            del_flag = True
         else:
-            print "Empty ID"
+            print "Argument provided is not a valid collection document id"
     except ValueError:
-        print "Could not remove from DB"
-        return Annotation.objects.all()
+        pass
 
-    print "Removed an Annotation"
-    return Annotation.objects.all()
+    if del_flag:
+        print "Removed an Annotation from DB"
+        return True
 
+    print "Could not remove from DB"
+    return False
 
 
 def CreateFromPOSTinfo( subject_url, object_json ):
@@ -56,7 +62,7 @@ def CreateFromPOSTinfo( subject_url, object_json ):
     except ValueError:
 
         print "Could not save to DB"
-        return Annotation.objects.all()
+        return False
 
     print "Created an Annotation"
-    return Annotation.objects.all()
+    return True
